Develop `\AdviceCsnameGetOriginal` and use it in auto-key `to context`.
`\AdviceCsnameGetOriginal` *accepts* a control sequence name (of the advised
command).  The previous attempt via `\AdviceGetOriginalCsname`,
which *produced* a control sequence name (of the control sequence holding the
original defitinion), is abandoned, and the macro deleted, because under that
approach, the advised control sequence (potentially internal) had to appear in
the c-memo.