Collargs' argument processors now work without a formal argument.
It was impossible (or at least too hard for me) to define expandable processors
and wraps taking a formal argument in Advice.  Advice stores the option
values (including `args` and `(raw) collector options`) in macros, whose
contents are passed around, so having `#1` in there was troublesome.  The new
argument processors expect to receive the input argument in token register
`\collargsArg` (where it always resided, anyway); this token register is now
both the input and the output of argument processors.

In practice, this means any `#1` in the definition of an argument processor
should be replaced by `\the\collargsArg`.

With this change, (pre/post)wraps become useless, as it is usually too
cumbersome to expand a definition containing `\the\collargsArg` in one step.
Keys `append postwrap` etc. are therefore removed.  However, for brevity,
expandable processors (which perform a full expansion) are now renamed to
wraps: `append expandable postprocessor` is renamed to `append postwrap`, etc.