Bugfix in `\AdviceGetOriginal`.
Replace `\noexpand` by `\unexpanded` in the "not-handled" branch of the
conditional, because the meaning of a `\noexpand`ed control sequence is
`\relax'.

(Remember that the entire complication, i.e. the reason we don't simply say
`\expandafter#2`, is that we want to produce `#2` under full expansion, as
advertised in the docs.)