Biblatex support: issue `\nocite` into cc-memo.
Originally, the cc-memo replicated the original command (I can't remember why).
Therefore, the cc-memo replica had to include all the arguments of the citation
command.  As the outer handler defined by auto-key `cite` grabs everything up
to the first braced token list, which contains the <key> argument, this made it
impossible to apply `cite` to `\citefield` and friends, where the <key>
argument is followed by other arguments.  Issuing `\nocite` into the cc-memo
avoids the problem, as this macro expects the <key> as the sole argument.

Minor changes:

* `\nocite` makes `\setbox0` unnecessary.

* `\mmz@biblatex@cite@do@keys` contains the code common to `cite` and `cites`.

* Allow parenthesized optional arguments in multicite commands.