Advice: adapt tracing to Collargs' `return`.
The assumption that the inner handler will always receive a mandatory argument
does not hold any more.  In production, this does not cause problems, as the
inner handler is task-specific and synchronized with the collector.  The
tracing inner handler, however, is generic.  The hack is to inspect
`\collargsReturn`.  Consequently, tracing collected arguments returned in a
form other than braced will only work with Collargs.  Perhaps foreign
collectors should be advised to set `\collargsReturn` if they return anything
else?