New initial value: `mkdir command` = `memoize-extract.pl --mkdir`.
In the case `mkdir command` remains unchanged by the user, `extract=python`
will reinitialize it to `memoize-extract.py --mkdir`.

`extract=no` leaves the initial value as it is, which is handy when triggering
extraction with `latexmk`.

`memoize-extract.pl/py` is in fact `\mmzvalueof{perl/python extraction
command}`, to automatically adapt to the user-given location of the script.