A new approach to the issue of the eof newline produced by `\scantokens`
The previous solution was to add `\endinput` at the end of the `\scantokens`
argument, but environment `Verbatim` of package Fancyvrb choked on this
`\endinput`.

I noticed the `\endinput` problem while working on issue #44
on (in)compatibility with package Minted, and this change is a part of the
solution for that issue. (The rest of the solution must be done from within
Minted, see Memoize's issue #44 for a link to the pull request.)