Implement `before/after [h/v]box` and `prepend/append to memoized code`
`before/after [h/v]box` are wrapped around `\mmzIncludeExtern` during
utilization, and around the typeset extern box after memoization.  They are
used instead of the recent solution, `typeset [h/v]box`, which turned out
cumbersome to use.

`prepend/append to memoized code` are similar to `at begin/end memoization`,
but the given code is processed by the `driver`.

Simplifying a recent solution for the extra space in the verbatim mode,
`\mmz@scantokens` does not depend on `capture` to set up the `\unskip` macro
for it any more.

Also implemented is a minor utility, `\mmzcmd`, and a `\noindentmacrocode` in
the code docs got a star argument, which reduces vertical space.